Folder refactoring: moved Migrations outside of Data and changed ApplicationDbContext's directory
