+++ MonkeyRadioStreaming/Startup.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using MonkeyRadioStreaming.Areas.Identity;
-using MonkeyRadioStreaming.Data;
+using MonkeyRadioStreaming.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
